Replace POINTS_PRECISION with global_config.points_dtype
Move the setting to `pyvista.global_config`, default it to `'preserve'`, and
enforce it centrally in `_update_alg`/`_get_output` instead of through a
per-filter `points_dtype` keyword.

+++ pyvista/examples/downloads.py
@@ -4147,7 +4147,7 @@ def _kitchen_split_load_func(mesh):  # noqa: ANN001
         alg.SetInputDataObject(mesh)
         alg.SetExtent(extent)  # type: ignore[call-overload]
         alg.Update()
-        result = _get_output(alg, points_dtype=np.single)
+        result = _get_output(alg)
         kitchen[key] = result
     return kitchen
